added some validation to the input. (Migrated by TFS Migration Toolkit) (Updated by TFS to TFS Connector, server: http://1x:8080/, id 1943. )
git-tfs-id: [https://tfs.caes.ucdavis.edu:8088/tfs/DefaultCollection]$/PTF/TRUNK/PTF;C1307

+++ PTF.Core/Domain/Experiment.cs
@@ -20,7 +20,7 @@ namespace CAESDO.PTF.Core.Domain
         [StringLengthValidator(50)]
         public virtual string Explant { get; set; }
         public virtual string Comments { get; set; }
-        public virtual float OpticalDensity { get; set; }
+        public virtual decimal OpticalDensity { get; set; }
 
         public virtual IList<Plant> Plants { get; set; }
         public virtual IList<Note> Notes { get; set; }
